requirements.txt
aiosmtplib==3.0.1
amqp==5.2.0
anyio==4.3.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
Cerberus==1.3.5
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-timezone-field==6.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
kombu==5.3.7
msgpack==1.0.8
numpy==1.26.4
pandas==2.2.1
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg2==2.9.9
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
toolz==0.12.1
typing_extensions==4.9.0
tzdata==2024.1
ulid-py==1.1.0
vine==5.1.0
wcwidth==0.2.13
